@@ -6,23 +6,32 @@
   "scripts": {
     "build": "tsc && tsc-alias && tsc-esm-fix",
     "start": "node dist/server.js",
-    "dev": "tsx watch src/server.ts"
+    "dev": "tsx watch src/server.ts",
+    "format": "prettier --write \"src/**/*.{js,ts,tsx,json,css,md}\"",
+    "lint": "eslint . --ext .ts,.js",
+    "lint:fix": "eslint . --ext .ts,.js --fix"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "type": "module",
   "devDependencies": {
+    "@eslint/js": "^9.39.1",
     "@types/cookie-parser": "^1.4.10",
     "@types/express": "^5.0.3",
     "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "^24.0.14",
+    "@types/pg": "^8.15.6",
+    "eslint": "^9.39.1",
+    "eslint-plugin": "^1.0.1",
     "nodemon": "^3.1.10",
+    "prettier": "^3.6.2",
     "ts-node": "^10.9.2",
     "tsc-alias": "^1.8.16",
     "tsc-esm-fix": "^3.1.2",
     "tsx": "^4.20.6",
-    "typescript": "^5.8.3"
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.46.4"
   },
   "dependencies": {
     "@grpc/grpc-js": "^1.13.4",
@@ -31,6 +40,8 @@
     "amqplib": "^0.10.8",
     "cookie-parser": "^1.4.7",
     "dotenv": "^17.2.3",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.4",
     "express": "^5.1.0",
     "fastify": "^5.6.2",
     "fastify-cookie": "^5.6.1",
@@ -38,6 +49,7 @@
     "inversify": "^7.10.4",
     "ioredis": "^5.8.2",
     "mongoose": "^8.16.3",
+    "pg": "^8.16.3",
     "pino-pretty": "^13.1.2",
     "reflect-metadata": "^0.2.2",
     "stripe": "^18.3.0",
