@@ -1,12 +1,15 @@
-import { sendUnaryData, ServerUnaryCall } from "@grpc/grpc-js";
-import { inject, injectable } from "inversify";
-import { TYPES } from "@/types/inversify-types";
-import { IPaymentService } from "@/services/interface/i-payment-service";
-import { IDriverWalletService } from "@/services/interface/i-driver-wallet-service";
+import { sendUnaryData, ServerUnaryCall } from '@grpc/grpc-js';
+import { inject, injectable } from 'inversify';
+import { TYPES } from '@/types/inversify-types';
+import { IPaymentService } from '@/services/interface/i-payment-service';
+import { IDriverWalletService } from '@/services/interface/i-driver-wallet-service';
 
 @injectable()
 export class GrpcPaymentController {
-  constructor(@inject(TYPES.DriverWalletService) private _driverWalletService:IDriverWalletService){}
+  constructor(
+    @inject(TYPES.DriverWalletService)
+    private _driverWalletService: IDriverWalletService
+  ) {}
 
   createDriverConnectAccount = async (
     call: ServerUnaryCall<
@@ -18,12 +21,15 @@ export class GrpcPaymentController {
     try {
       const { driverId, email } = call.request;
 
-      const accountData = await this._driverWalletService.createDriverConnectAccount( email,driverId );
+      const accountData = await this._driverWalletService.createDriverConnectAccount(
+        email,
+        driverId
+      );
 
       callback(null, accountData);
     } catch (err) {
       console.log(err);
-      throw new Error("Stripe account creation failed");
+      throw new Error('Stripe account creation failed');
     }
   };
 }
