@@ -1,12 +1,12 @@
-import { sendUnaryData, ServerUnaryCall } from "@grpc/grpc-js";
-import { ConformCashPaymentDto } from "../dto/paymentRes.dto";
-import { IPaymentService } from "../services/interface/i-payment-service";
-import { PaymentReq } from "../types/request";
-import { IncomingHttpHeaders } from "http";
-import { IStripeService } from "../services/interface/i-stripe-service";
-import { InternalError, IResponse } from "@Pick2Me/shared";
-import { inject, injectable } from "inversify";
-import { TYPES } from "@/types/inversify-types";
+import { sendUnaryData, ServerUnaryCall } from '@grpc/grpc-js';
+import { ConformCashPaymentDto } from '../dto/paymentRes.dto';
+import { IPaymentService } from '../services/interface/i-payment-service';
+import { PaymentReq } from '../types/request';
+import { IncomingHttpHeaders } from 'http';
+import { IStripeService } from '../services/interface/i-stripe-service';
+import { InternalError, IResponse } from '@Pick2Me/shared';
+import { inject, injectable } from 'inversify';
+import { TYPES } from '@/types/inversify-types';
 
 @injectable()
 export class PaymentController {
@@ -20,27 +20,22 @@ export class PaymentController {
     callback: (error: Error | null, response: any) => void
   ) {
     try {
-      const result = await this._stripeService.createCheckoutSession(
-        call.request
-      );
+      const result = await this._stripeService.createCheckoutSession(call.request);
       callback(null, result);
     } catch (error) {
       InternalError(error, callback);
     }
   }
 
-  async handleStripeWebhook(
-    rawBody: Buffer,
-    headers: IncomingHttpHeaders
-  ): Promise<void> {
+  async handleStripeWebhook(rawBody: Buffer, headers: IncomingHttpHeaders): Promise<void> {
     if (!rawBody || !headers) {
-      throw new Error("Missing webhook payload or headers");
+      throw new Error('Missing webhook payload or headers');
     }
 
     try {
       await this._stripeService.handleStripeWebhook(rawBody, headers);
     } catch (err: any) {
-      console.error("PaymentController.handleStripeWebhook error", {
+      console.error('PaymentController.handleStripeWebhook error', {
         error: err?.message ?? err,
       });
       throw err;
@@ -61,9 +56,7 @@ export class PaymentController {
     callback: sendUnaryData<IResponse<ConformCashPaymentDto>>
   ) {
     try {
-      const result = await this._paymentService.ConfirmCashPayment(
-        call.request
-      );
+      const result = await this._paymentService.ConfirmCashPayment(call.request);
 
       callback(null, result);
     } catch (error) {
