@@ -1,14 +1,14 @@
-import { bookingProto,driverProto } from "@Pick2Me/shared";
-import * as grpc from "@grpc/grpc-js";
+import { bookingProto, driverProto } from '@Pick2Me/shared';
+import * as grpc from '@grpc/grpc-js';
 
 const driverClient = new (driverProto as any).Driver(
-    process.env.DRIVER_GRPC_URL,
-    grpc.credentials.createInsecure()
+  process.env.DRIVER_GRPC_URL,
+  grpc.credentials.createInsecure()
 );
 
-const bookingClient = new (bookingProto as any).Booking( 
-    process.env.BOOKING_GRPC_URL,
-    grpc.credentials.createInsecure()
+const bookingClient = new (bookingProto as any).Booking(
+  process.env.BOOKING_GRPC_URL,
+  grpc.credentials.createInsecure()
 );
 
-export { driverClient,  bookingClient};
+export { driverClient, bookingClient };
