@@ -1,13 +1,12 @@
-import { GrpcPaymentController } from "../../controllers/grpc-payment-controller";
-
+import { GrpcPaymentController } from '../../controllers/grpc-payment-controller';
 
 type Handlers = {
-  grpcPaymentController: GrpcPaymentController
+  grpcPaymentController: GrpcPaymentController;
 };
 
-export function createPaymentHandlers(controller:Handlers){
-    const {grpcPaymentController} = controller;
-    return{
-        CreateDriverConnectAccount: grpcPaymentController.createDriverConnectAccount
-    }
+export function createPaymentHandlers(controller: Handlers) {
+  const { grpcPaymentController } = controller;
+  return {
+    CreateDriverConnectAccount: grpcPaymentController.createDriverConnectAccount,
+  };
 }
