@@ -1,7 +1,7 @@
 import mongoose, { Schema, Document } from 'mongoose';
 
 export interface ITransaction extends Document {
-  _id: mongoose.ObjectId,
+  _id: mongoose.ObjectId;
   transactionId: string;
   bookingId: string;
   userId: string;
@@ -25,8 +25,16 @@ const transactionSchema = new Schema<ITransaction>(
     userId: { type: String, required: true },
     driverId: { type: String, required: true },
     amount: { type: Number, required: true },
-    paymentMethod: { type: String, enum: ['stripe', 'wallet', 'cash'], required: true },
-    status: { type: String, enum: ['pending', 'completed', 'failed'], required: true },
+    paymentMethod: {
+      type: String,
+      enum: ['stripe', 'wallet', 'cash'],
+      required: true,
+    },
+    status: {
+      type: String,
+      enum: ['pending', 'completed', 'failed'],
+      required: true,
+    },
     adminShare: { type: Number, required: true },
     driverShare: { type: Number, required: true },
     stripeSessionId: { type: String },
@@ -40,9 +48,7 @@ const transactionSchema = new Schema<ITransaction>(
 
 export const TransactionModel = mongoose.model<ITransaction>('Transaction', transactionSchema);
 
-
-// The Guy Who Froze at One English Sentence… Now Teaches DSA 
-
+// The Guy Who Froze at One English Sentence… Now Teaches DSA
 
 // There was a guy who couldn’t even write a proper English sentence.
 // No big goals. No direction. Just vibes.
@@ -51,11 +57,11 @@ export const TransactionModel = mongoose.model<ITransaction>('Transaction', tran
 // After 10th grade, he joined a private college.
 // Why private? Because no government college wanted a student with zero A+.
 
-// He tried to take Commerce, but froze when the teacher asked to write one English sentence. Just one. 
+// He tried to take Commerce, but froze when the teacher asked to write one English sentence. Just one.
 
-// The teacher smiled, “Take Humanities.” 
+// The teacher smiled, “Take Humanities.”
 // But our hero said, “No sir, Commerce.”
-// Not because he loved accounts, just because his friends were there. 
+// Not because he loved accounts, just because his friends were there.
 // Comfort zone > interest.
 
 // Somehow, he completed +2 and joined BCom.
@@ -65,13 +71,13 @@ export const TransactionModel = mongoose.model<ITransaction>('Transaction', tran
 // Comfort zone wins again.
 
 // Then one random YouTube video changed everything.
-// Nikhil Kilivayil from Brototype said anyone can become a software developer with basic maths and English. 
+// Nikhil Kilivayil from Brototype said anyone can become a software developer with basic maths and English.
 
 // He started learning to code through YouTube.
-// And surprisingly, he loved it — debugging, solving problems, building things. 
+// And surprisingly, he loved it — debugging, solving problems, building things.
 
 // Plot twist part two: basic English wasn’t enough.
-// So he learned English the same way he learned coding, through YouTube. 
+// So he learned English the same way he learned coding, through YouTube.
 // Maybe all that debugging fixed his brain too.
 // Because suddenly, things started making sense.
 // And for once, he didn’t run away.
