@@ -1,12 +1,12 @@
-import fp from "fastify-plugin";
-import { FastifyPluginAsync } from "fastify";
-import { HttpError } from "@Pick2Me/shared";
+import fp from 'fastify-plugin';
+import { FastifyPluginAsync } from 'fastify';
+import { HttpError } from '@Pick2Me/shared';
 
 const errorHandlerPlugin: FastifyPluginAsync = async (fastify) => {
   fastify.setErrorHandler((err, request, reply) => {
     if (err instanceof HttpError) {
       const payload = {
-        message: err.message || "Error",
+        message: err.message || 'Error',
         code: err.code || undefined,
       };
       reply.status(err.status).send(payload);
@@ -14,8 +14,8 @@ const errorHandlerPlugin: FastifyPluginAsync = async (fastify) => {
     }
 
     fastify.log.error(err);
-    reply.status(500).send({ message: "Internal server error" });
+    reply.status(500).send({ message: 'Internal server error' });
   });
 };
 
-export default fp(errorHandlerPlugin, { name: "errorHandlerPlugin" });
+export default fp(errorHandlerPlugin, { name: 'errorHandlerPlugin' });
