@@ -1,19 +1,15 @@
-import { Wallet } from "@/models/wallet.entity";
-import { QueryRunner } from "typeorm";
+import { Wallet } from '@/models/wallet.entity';
+import { QueryRunner } from 'typeorm';
 
 export interface IWalletRepository {
-  getForUpdate(
-    queryRunner: QueryRunner,
-    userId: string,
-    currency: string
-  ): Promise<Wallet | null>;
+  getForUpdate(queryRunner: QueryRunner, userId: string, currency: string): Promise<Wallet | null>;
   createIfNotExists(userId: string, currency: string);
   applyTransactionTransactional(params: {
     queryRunner: QueryRunner;
     userId: string;
     currency?: string;
     amount: bigint;
-    direction: "credit" | "debit";
+    direction: 'credit' | 'debit';
     reason?: string;
     referenceType?: string;
     referenceId?: string;
@@ -26,7 +22,7 @@ export interface IWalletRepository {
     userId: string;
     currency?: string;
     amount: bigint;
-    direction: "credit" | "debit";
+    direction: 'credit' | 'debit';
     reason?: string;
     referenceType?: string;
     referenceId?: string;
