@@ -5,7 +5,12 @@ export interface ITransactionRepository {
   findByTransactionId(transactionId: string): Promise<ITransaction | null>;
   findByIdempotencyKey(idempotencyKey: string): Promise<ITransaction | null>;
   update(transactionId: string, update: Partial<ITransaction>): Promise<ITransaction | null>;
-  updateStatusByKey(idempotencyKey: string, status: 'pending' | 'completed' | 'failed'): Promise<ITransaction | null>
-  updateStatus(transactionId: string, status: 'pending' | 'completed' | 'failed'): Promise<ITransaction | null>
+  updateStatusByKey(
+    idempotencyKey: string,
+    status: 'pending' | 'completed' | 'failed'
+  ): Promise<ITransaction | null>;
+  updateStatus(
+    transactionId: string,
+    status: 'pending' | 'completed' | 'failed'
+  ): Promise<ITransaction | null>;
 }
-
