@@ -1,5 +1,6 @@
-
-
-export interface IDriverWalletService{
-    createDriverConnectAccount(email: string, driverId: string): Promise<{ accountId: string; accountLinkUrl: string }>
+export interface IDriverWalletService {
+  createDriverConnectAccount(
+    email: string,
+    driverId: string
+  ): Promise<{ accountId: string; accountLinkUrl: string }>;
 }
