@@ -1,7 +1,7 @@
-import { IncomingHttpHeaders } from "http";
-import { ConformCashPaymentDto } from "../../dto/paymentRes.dto";
-import { PaymentReq } from "../../types/request";
-import { StripeCheckoutSessionRes } from "../../types/response";
+import { IncomingHttpHeaders } from 'http';
+import { ConformCashPaymentDto } from '../../dto/paymentRes.dto';
+import { PaymentReq } from '../../types/request';
+import { StripeCheckoutSessionRes } from '../../types/response';
 
 export interface IPaymentService {
   ConfirmCashPayment(data: {
