@@ -1,5 +1,3 @@
-
-
-export interface IUserWalletService{
-    createWalletForUser(userId:string,currency?:string): Promise<void>;
+export interface IUserWalletService {
+  createWalletForUser(userId: string, currency?: string): Promise<void>;
 }
