@@ -1,9 +1,6 @@
-
-
-
-export interface PaymentReq{
-    bookingId:string,
-    userId: string,
-    driverId: string,
-    amount: number
+export interface PaymentReq {
+  bookingId: string;
+  userId: string;
+  driverId: string;
+  amount: number;
 }
