@@ -1,11 +1,8 @@
 import { FastifyInstance } from 'fastify';
 import rawbody from '../plugins/rawbody';
 
-
 export function StripeWebhook(fastify: FastifyInstance) {
   fastify.register(rawbody);
 
-  fastify.post('/webhook/stripe',()=>{
-    
-  });
+  fastify.post('/webhook/stripe', () => {});
 }
